requirements.txt
schedule==1.2.*
python-dotenv==1.0.*
pytz==2023.3